requirements.txt
streamlit
pandas
matplotlib
seaborn
fpdf2